changing skillid to uint16

+++ src/map/clif.hpp
@@ -930,7 +930,7 @@ void clif_class_change_target(struct block_list *bl,int class_, int type, enum s
 
 void clif_skillinfoblock(map_session_data *sd);
 void clif_skillup(map_session_data *sd, uint16 skill_id, int lv, int range, int upgradable);
-void clif_skillinfo( map_session_data &sd, int32 skill_id );
+void clif_skillinfo( map_session_data &sd, uint16 skill_id );
 void clif_addskill(map_session_data *sd, int skill_id);
 void clif_deleteskill(map_session_data *sd, int skill_id, bool skip_infoblock = false);
